KAFKA-10070: parameterize Connect unit tests to remove code duplication
